Add node toggle to publish docs github action
I suggest that we do not use the full package.json install to build the
documentation if the application does not need any yarn dependencies for
the dox specs to pass. The package.json should contain runtime
depencencies and development dependencies, and arguably the aglio or
redoc-cli used to build the documentation fits in neither of those
groups.
Added a use_node input that defaults to true, similar to how it is done
in the build workflow. This enables us to use node for runtime
dependencies needed by the dox specs to successfully run. Added global
install of redoc-cli to the publish_docs script. This way the
publish_docs github action does not need to install all development
dependencies, making it faster.