Remove build and deploy workflows for docker apps
When using docker to deploy an application we do not need the mina gem
or the deploy scripts and workflows.
Reorganized the template file a bit so that we first gather most of the
user input and only then install the bundle, before running the last
scripts for the setup process.